refactor(gateway): move broker commands into own package
