+++ gateway/src/main/java/io/camunda/zeebe/gateway/RequestMapper.java
@@ -10,18 +10,18 @@ package io.camunda.zeebe.gateway;
 import static org.agrona.LangUtil.rethrowUnchecked;
 
 import com.fasterxml.jackson.core.JsonParseException;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerActivateJobsRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerCancelProcessInstanceRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerCompleteJobRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerCreateProcessInstanceRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerCreateProcessInstanceWithResultRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerDeployProcessRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerFailJobRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerPublishMessageRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerResolveIncidentRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerSetVariablesRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerThrowErrorRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerUpdateJobRetriesRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerActivateJobsRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerCancelProcessInstanceRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerCompleteJobRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerCreateProcessInstanceRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerCreateProcessInstanceWithResultRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerDeployProcessRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerFailJobRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerPublishMessageRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerResolveIncidentRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerSetVariablesRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerThrowErrorRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerUpdateJobRetriesRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.ActivateJobsRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.CancelProcessInstanceRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.CompleteJobRequest;
