+++ gateway/src/main/java/io/camunda/zeebe/gateway/impl/broker/BrokerRequestManager.java
@@ -17,8 +17,8 @@ import io.camunda.zeebe.gateway.cmd.PartitionNotFoundException;
 import io.camunda.zeebe.gateway.impl.ErrorResponseHandler;
 import io.camunda.zeebe.gateway.impl.broker.cluster.BrokerClusterState;
 import io.camunda.zeebe.gateway.impl.broker.cluster.BrokerTopologyManagerImpl;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerPublishMessageRequest;
 import io.camunda.zeebe.gateway.impl.broker.request.BrokerRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerPublishMessageRequest;
 import io.camunda.zeebe.gateway.impl.broker.response.BrokerResponse;
 import io.camunda.zeebe.gateway.metrics.GatewayMetrics;
 import io.camunda.zeebe.protocol.Protocol;
