+++ gateway/src/test/java/io/camunda/zeebe/gateway/api/job/CompleteJobStub.java
@@ -9,7 +9,7 @@ package io.camunda.zeebe.gateway.api.job;
 
 import io.camunda.zeebe.gateway.api.util.StubbedBrokerClient;
 import io.camunda.zeebe.gateway.api.util.StubbedBrokerClient.RequestStub;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerCompleteJobRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerCompleteJobRequest;
 import io.camunda.zeebe.gateway.impl.broker.response.BrokerResponse;
 import io.camunda.zeebe.protocol.impl.record.value.job.JobRecord;
