+++ gateway/src/test/java/io/camunda/zeebe/gateway/api/job/UpdateJobRetriesTest.java
@@ -10,7 +10,7 @@ package io.camunda.zeebe.gateway.api.job;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.zeebe.gateway.api.util.GatewayTest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerUpdateJobRetriesRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerUpdateJobRetriesRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.UpdateJobRetriesRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.UpdateJobRetriesResponse;
 import io.camunda.zeebe.protocol.impl.record.value.job.JobRecord;
