+++ gateway/src/test/java/io/camunda/zeebe/gateway/api/process/CancelProcessInstanceStub.java
@@ -9,7 +9,7 @@ package io.camunda.zeebe.gateway.api.process;
 
 import io.camunda.zeebe.gateway.api.util.StubbedBrokerClient;
 import io.camunda.zeebe.gateway.api.util.StubbedBrokerClient.RequestStub;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerCancelProcessInstanceRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerCancelProcessInstanceRequest;
 import io.camunda.zeebe.gateway.impl.broker.response.BrokerResponse;
 import io.camunda.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceRecord;
