+++ gateway/src/test/java/io/camunda/zeebe/gateway/api/process/PublishMessageStub.java
@@ -9,7 +9,7 @@ package io.camunda.zeebe.gateway.api.process;
 
 import io.camunda.zeebe.gateway.api.util.StubbedBrokerClient;
 import io.camunda.zeebe.gateway.api.util.StubbedBrokerClient.RequestStub;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerPublishMessageRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerPublishMessageRequest;
 import io.camunda.zeebe.gateway.impl.broker.response.BrokerResponse;
 
 public final class PublishMessageStub
