+++ gateway/src/test/java/io/camunda/zeebe/gateway/api/process/ResolveIncidentTest.java
@@ -10,7 +10,7 @@ package io.camunda.zeebe.gateway.api.process;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.zeebe.gateway.api.util.GatewayTest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerResolveIncidentRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerResolveIncidentRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.ResolveIncidentRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.ResolveIncidentResponse;
 import io.camunda.zeebe.protocol.record.ValueType;
