+++ gateway/src/test/java/io/camunda/zeebe/gateway/api/process/SetVariablesTest.java
@@ -10,7 +10,7 @@ package io.camunda.zeebe.gateway.api.process;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.zeebe.gateway.api.util.GatewayTest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerSetVariablesRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.command.BrokerSetVariablesRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.SetVariablesRequest;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.SetVariablesResponse;
 import io.camunda.zeebe.protocol.Protocol;
